split test_core module into per-model tests
This module was getting a little large, so this is a pure location
refactor to ease finding tests.
