+++ tests/unit/jobserver/models/test_project.py
@@ -1,0 +1,98 @@
+from django.urls import reverse
+
+from ....factories import OrgFactory, ProjectFactory
+
+
+def test_project_get_absolute_url():
+    org = OrgFactory(name="test-org")
+    project = ProjectFactory(org=org)
+    url = project.get_absolute_url()
+    assert url == reverse(
+        "project-detail", kwargs={"org_slug": org.slug, "project_slug": project.slug}
+    )
+
+
+def test_project_get_approved_url_with_number():
+    project = ProjectFactory(number=42)
+
+    assert (
+        project.get_approved_url()
+        == "https://www.opensafely.org/approved-projects#project-42"
+    )
+
+
+def test_project_get_approved_url_without_number():
+    project = ProjectFactory()
+
+    assert project.get_approved_url() == "https://www.opensafely.org/approved-projects"
+
+
+def test_project_get_edit_url():
+    project = ProjectFactory()
+
+    url = project.get_edit_url()
+
+    assert url == reverse(
+        "project-edit",
+        kwargs={"org_slug": project.org.slug, "project_slug": project.slug},
+    )
+
+
+def test_project_get_staff_edit_url():
+    project = ProjectFactory()
+
+    url = project.get_staff_edit_url()
+
+    assert url == reverse("staff:project-edit", kwargs={"slug": project.slug})
+
+
+def test_project_get_releases_url():
+    project = ProjectFactory()
+
+    url = project.get_releases_url()
+
+    assert url == reverse(
+        "project-release-list",
+        kwargs={
+            "org_slug": project.org.slug,
+            "project_slug": project.slug,
+        },
+    )
+
+
+def test_project_get_staff_url():
+    project = ProjectFactory()
+
+    url = project.get_staff_url()
+
+    assert url == reverse("staff:project-detail", kwargs={"slug": project.slug})
+
+
+def test_project_get_staff_feature_flags_url():
+    project = ProjectFactory()
+
+    url = project.get_staff_feature_flags_url()
+
+    assert url == reverse("staff:project-feature-flags", kwargs={"slug": project.slug})
+
+
+def test_project_populates_slug():
+    assert ProjectFactory(name="Test Project", slug="").slug == "test-project"
+
+
+def test_project_str():
+    org = OrgFactory(name="test-org")
+
+    project = ProjectFactory(org=org, name="Very Good Project")
+    assert str(project) == "test-org | Very Good Project"
+
+    project = ProjectFactory(org=org, name="Another Very Good Project", number=42)
+    assert str(project) == "test-org | 42 - Another Very Good Project"
+
+
+def test_project_title():
+    project = ProjectFactory(number=None)
+    assert project.title == project.name
+
+    project = ProjectFactory(name="test", number=123)
+    assert project.title == "123 - test"
